lists file amendment

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,15 +1,15 @@
 #include "lists.h"
+#include <stdio.h>
 
 /**
- * print_listint - function with one argument
- * @h: pointer to struct
+ * print_listint -  prints all the elements of a listint_t list.
+ * @h: list to print
  *
- * Description: prints all the elements of a list
- * Return: number of nodes
+ * Return: number of elements in the list
  */
 size_t print_listint(const listint_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h != NULL)
 	{
